+++ error/index.js
@@ -1,10 +1,12 @@
 'use strict';
 
 const BusinessLogicError = require('./business-logic-error/business-logic-error');
+const NotFoundError = require('./not-found-error/not-found-error');
 const errorType = require('./type/error-type');
 
 module.exports = {
     BusinessLogicError,
-
+    NotFoundError,
+    
     errorType,
 };