+++ error/type/error-type.js
@@ -1,5 +1,9 @@
 'use strict';
 
+/** @typedef {string} ErrorId */
+/** @typedef {Object<string,ErrorId>} ErrorType */
+
 module.exports = {
-    
+    UnknownManager: '783b3f7e-5014-4efb-b2ae-6a4221d97f8f',
+    NoManagerSecret: 'cbdf185a-3e5b-45de-91a5-5a3c09b66be4',
 };