+++ service/team/team-service.js
@@ -1,0 +1,14 @@
+'use strict';
+
+const { Logable } = require('../../lib/log');
+
+class TeamService extends Logable {
+    
+    constructor(options) {
+        super(options);
+
+        this._teamModel 
+    }
+}
+
+module.exports = TeamService;