+++ transport/bot/controller/command/create-team/create-team-command-controller.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const { Controller } = require('../../../../../lib/controller');
+
+class CreateTeamCommandController extends Controller {
+
+}
+
+module.exports = CreateTeamCommandController;