+++ transport/bot/controller/command/create-transfer/create-transfer-command-controller.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const { Controller } = require('../../../../../lib/controller');
+
+class CreateTransferCommandController extends Controller {
+
+}
+
+module.exports = CreateTransferCommandController;