+++ transport/bot/controller/command/delete-transfer/delete-transfer-command-controller.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const { Controller } = require('../../../../../lib/controller');
+
+class DeleteTransferCommandController extends Controller {
+
+}
+
+module.exports = DeleteTransferCommandController;