+++ transport/bot/controller/command/get-transfer/get-transfer-command-controller.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const { Controller } = require('../../../../../lib/controller');
+
+class GetTransferCommandController extends Controller {
+
+}
+
+module.exports = GetTransferCommandController;