+++ transport/bot/controller/command/index.js
@@ -1,13 +1,29 @@
 'use strict';
 
-const CreateCommandController = require('./create/create-command-controller');
 const InfoCommandController = require('./info/info-command-controller');
-const JoinCommandController = require('./join/join-command-controller');
-const LeaveCommandController = require('./leave/leave-command-controller');
+const StartCommandController = require('./start/start-command-controller');
+
+const CreateLeagueCommandController = require('./create-league/create-league-command-controller');
+const JoinLeagueCommandController = require('./join-league/join-league-command-controller');
+const LeaveLeagueCommandController = require('./leave-league/leave-league-command-controller');
+
+const CreateTeamCommandController = require('./create-team/create-team-command-controller');
+
+const CreateTransferCommandController = require('./create-transfer/create-transfer-command-controller');
+const GetTransferCommandController = require('./get-transfer/get-transfer-command-controller');
+const DeleteTransferCommandController = require('./delete-transfer/delete-transfer-command-controller');
 
 module.exports = {
-    CreateCommandController,
     InfoCommandController,
-    JoinCommandController,
-    LeaveCommandController,
+    StartCommandController,
+
+    CreateLeagueCommandController,
+    JoinLeagueCommandController,
+    LeaveLeagueCommandController,
+
+    CreateTeamCommandController,
+
+    CreateTransferCommandController,
+    GetTransferCommandController,
+    DeleteTransferCommandController,
 };