+++ transport/bot/controller/command/info/info-command-controller.js
@@ -2,12 +2,22 @@
 
 const { Controller } = require('../../../../../lib/controller');
 
+/**
+ * @typedef {Object} InfoCommandControllerOptions
+ * @property {import('../../../../../service/league/league-service')} leagueService -
+ * @property {import('../../../../../service/manager/manager-service')} managerService -
+ */
+/** @typedef {InfoCommandControllerOptions & import('../../../../../lib/controller/controller').Options} Options */
+
 class InfoCommandController extends Controller {
 
+    /**
+     * @constructor
+     * @param {Options} options -
+     */
     constructor(options) {
         super(options);
 
-        this._command = '/info';
         this._leagueService = options.leagueService;
         this._managerService = options.managerService;
     }
@@ -21,25 +31,23 @@ class InfoCommandController extends Controller {
 
         const {
             userId,
+            firstName: name,
         } = manager;
 
         this.log(ctx, 'Getting manager info.', { userId });
 
-        const {
-            name,
-        } = manager;
-
         let reply = `Hello, ${name}!`;
 
         if (this._managerService.hasLeague(ctx, { manager })) {
-            const secret = await this._managerService.getSecret(ctx, { manager });
+            const leagueId = await this._managerService.getLeagueId(ctx, { manager });
             const {
-                name
-            } = await this._leagueService.getBySecret(ctx, { secret });
+                name: leagueName,
+                secret,
+            } = await this._leagueService.getByLeagueId(ctx, { leagueId });
 
-            reply = `${reply}\nYou are in a ${name} league. Your league secret is ${secret}.`;
+            reply = `${reply}\nYou are in a ${leagueName} league. Your league secret is ${secret}.`;
         } else {
-            reply = `${reply}\nYou have not joined a league yet. To join a league use /create [name] or /join [secret] commands.`;
+            reply = `${reply}\nYou have not joined a league yet. To join a league use /createleague [name] or /joinleague [secret] commands.`;
         }
 
         request.reply(reply);