+++ transport/bot/controller/command/join-league/join-league-command-controller.js
@@ -7,7 +7,6 @@ class JoinCommandController extends Controller {
     constructor(options) {
         super(options);
 
-        this._command = '/join';
         this._leagueService = options.leagueService;
         this._managerService = options.managerService;
     }
@@ -26,18 +25,18 @@ class JoinCommandController extends Controller {
             userId,
         } = manager;
 
-        const secret = text.substr(this._command.length + 1);
+        const secret = this._getText(ctx, { request });
 
         if (!secret) {
             this.logWarn(ctx, 'Unable to join a user to a league without a secret.');
-            request.reply('Please, sent a league secret. Use /join [secret] command.');
+            request.reply('Please, sent a league secret. Use /join-league [secret] command.');
 
             return;
         }
 
         if (this._managerService.hasLeague(ctx, { manager })) {
             request.reply('You have already in a league.');
-            request.reply('Leave a league with /leave command first.');
+            request.reply('Leave a league with /leave-league command first.');
 
             return;
         }
@@ -48,16 +47,18 @@ class JoinCommandController extends Controller {
 
         if (!league) {
             request.reply('A league with secret ' + secret + ' is not found.');
-            request.reply('Create a league with /create [name] command or use another secret.');
+            request.reply('Create a league with /create-league [name] command or use another secret.');
             
             return;
         }
 
-        await this._managerService.joinLeague(ctx, { userId, secret });
+        const { leagueId, name } = league;
 
-        request.reply('You have joined a league.');
+        await this._managerService.joinLeague(ctx, { userId, leagueId });
 
-        this.log(ctx, 'A user joined to a league.', { userId, secret });
+        request.reply(`You have joined a league [${name}].`);
+
+        this.log(ctx, 'A user joined to a league.', { userId, leagueId });
     }
 }